feat: 상품 상세 정보 api 연동

+++ app/src/main/java/woowacourse/shopping/data/product/remote/dto/ProductResponseDto.kt
@@ -19,7 +19,7 @@ data class ProductResponseDto(
 ) {
     fun toDomain(): Product =
         Product(
-            id = id.toLong(),
+            id = id,
             name = name,
             price = price,
             imageUrl = imageUrl,
